Create QuickActions component
